requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Machine learning libraries
xgboost>=1.6.0
tensorflow>=2.10.0
keras>=2.10.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
dash>=2.6.0
dash-bootstrap-components>=1.2.0

# Data processing and web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Geographic data processing
geopandas>=0.11.0
folium>=0.12.0

# Progress bars and utilities
tqdm>=4.64.0
joblib>=1.1.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0

# Statistical analysis
statsmodels>=0.13.0

# Model interpretation
shap>=0.41.0
lime>=0.2.0

# Data validation
great-expectations>=0.15.0

# Configuration management
pyyaml>=6.0
python-dotenv>=0.19.0

# Web framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# Additional dependencies for web app
Werkzeug>=2.3.0
Jinja2>=3.1.0
